T7.X (#308)

+++ src/main/java/com/gtnewhorizons/galaxia/registry/celestial/station/Station.java
@@ -55,7 +55,7 @@ public class Station extends CelestialAsset {
     }
 
     @Override
-    public long updateContents(InventoryKey item, int delta, boolean sync) {
+    public long updateContents(InventoryKey item, long delta, boolean sync) {
         return updateContents(item, delta);
     }
